+++ lib/workling/clients/amqp_client.rb
@@ -7,31 +7,23 @@ require 'mq'
 module Workling
   module Clients
     class AmqpClient < Workling::Clients::Base
-
+      
       # connects to the queue server
-      def connect
-        @amq = MQ.new
-      end
+      def connect; @amq = MQ.new; end
       
       # disconnect from the queue server
-      def close
-        @amq.close
-      end
-
+      def close; @amq.close; end
+      
       # subscribe to a queue
       def subscribe(key)
         @amq.queue(key).subscribe do |value|
           yield value
         end
       end
-
-      def get(key)
-        @amq.queue(key)
-      end
-
-      def set(key, value)
-        @amq.queue(key).publish(value)
-      end
+      
+      # request and retrieve work
+      def retrieve(key); @amq.queue(key); end
+      def request(key, value); @amq.queue(key).publish(value); end
     end
   end
 end