+++ lib/workling/remote.rb
@@ -16,7 +16,7 @@ module Workling
     
     # set the desired invoker. this class grabs work from the job broker and executes it. 
     mattr_accessor :invoker
-    @@invoker = Workling::Remote::Invokers::Poller
+    @@invoker = Workling::Remote::Invokers::ThreadedPoller
     
     # retrieve the dispatcher or instantiate it using the defaults
     def self.dispatcher