+++ script/listen.rb
@@ -3,19 +3,19 @@ puts '=> Loading Rails...'
 require File.dirname(__FILE__) + '/../../../../config/environment'
 require File.dirname(__FILE__) + '/../lib/workling/remote'
 require File.dirname(__FILE__) + '/../lib/workling/remote/invokers/poller'
-require File.dirname(__FILE__) + '/../lib/workling/remote/invokers/subscriber'
+require File.dirname(__FILE__) + '/../lib/workling/remote/invokers/eventmachine_subscriber'
 require File.dirname(__FILE__) + '/../lib/workling/routing/class_and_method_routing'
-
-puts '** Rails loaded.'
-puts '** Starting Workling::Remote::Invokers::Poller...'
-puts '** Use CTRL-C to stop.'
-
-ActiveRecord::Base.logger = Workling::Base.logger
-ActionController::Base.logger = Workling::Base.logger
 
 client = Workling::Remote.dispatcher.client
 invoker = Workling::Remote.invoker
 poller = invoker.new(Workling::Routing::ClassAndMethodRouting.new, client.class)
+
+puts '** Rails loaded.'
+puts "** Starting #{ invoker }..."
+puts '** Use CTRL-C to stop.'
+
+ActiveRecord::Base.logger = Workling::Base.logger
+ActionController::Base.logger = Workling::Base.logger
 
 trap(:INT) { poller.stop; exit }