+++ test/memcachequeue_client_test.rb
@@ -3,7 +3,7 @@ require File.dirname(__FILE__) + '/test_helper'
 context "The memcachequeue client" do
   specify "should be able to connect to multiple queue instances" do
     Workling.send :class_variable_set, "@@config", { :listens_on => "localhost:12345, 127.0.0.1:12346", :memcache_options => { :namespace => "myapp_development" } }
-    client = Workling::Clients::MemcacheQueue.new
+    client = Workling::Clients::MemcacheQueueClient.new
     client.connect
     
     client.queueserver_urls.should.equal ["localhost:12345", "127.0.0.1:12346"]
@@ -16,7 +16,7 @@ context "The memcachequeue client" do
   
   specify "should load it's config as well as any given MemCache options from RAILS_ENV/config/workling.yml" do
     Workling.send :class_variable_set, "@@config", { :listens_on => "localhost:12345", :memcache_options => { :namespace => "myapp_development" } }
-    client = Workling::Clients::MemcacheQueue.new
+    client = Workling::Clients::MemcacheQueueClient.new
     client.connect
     
     client.queueserver_urls.should.equal ["localhost:12345"]
@@ -28,7 +28,7 @@ context "The memcachequeue client" do
   
   specify "should load it's config correctly if no memcache options are given" do
     Workling.send :class_variable_set, "@@config", { :listens_on => "localhost:12345" }
-    client = Workling::Clients::MemcacheQueue.new
+    client = Workling::Clients::MemcacheQueueClient.new
     client.connect
 
     client.queueserver_urls.should.equal ["localhost:12345"]