+++ test/mocks/client.rb
@@ -2,7 +2,7 @@ require 'workling/clients/base'
 
 module Workling
   module Clients
-    class MemcacheQueue < Workling::Clients::Base
+    class MemcacheQueueClient < Workling::Clients::Base
       def raise_unless_connected!; end
     end
   end